requirements.txt
streamlit
pillow
transformers
fpdf
